requirements.txt
mkl
tensorboard
matplotlib
git+https://github.com/tinyobjloader/tinyobjloader.git@v2.0.0rc8#subdirectory=python
pybind11
pyyaml
pandas
tqdm
Pillow
numpy~=1.23.0
scipy>=1.7.2
scikit-image
six>=1.12.0
moviepy
opencv-python
protobuf>=3.20.0
polyscope
more_itertools
Futureproof
IPython
pyarrow
fastparquet
PyDispatcher
pynvml
setuptools==59.5.0
wandb>=0.13.5
pytest>=7.1.0
attrdict
einops
torchac
# pyexr